Migrate fixture to conftest.py
