+++ tests/ode/test_files.py
@@ -1,16 +1,7 @@
 import json
 import pytest
-import ode
 
 from ode.file import File
-
-
-@pytest.fixture(autouse=True)
-def project_folder(tmp_path):
-    # Patch the PROJECT_PATH to use temporary directory
-    ode.paths.PROJECT_PATH = tmp_path
-    ode.paths.METADATA_PATH = (tmp_path / ".metadata")
-    return tmp_path
 
 
 class TestFiles():
