+++ tests/ode/test_paths.py
@@ -1,15 +1,4 @@
-import pytest
-
-import ode
 from ode.paths import Paths
-
-
-@pytest.fixture(autouse=True)
-def project_folder(tmp_path):
-    # Patch the PROJECT_PATH to use temporary directory
-    ode.paths.PROJECT_PATH = tmp_path
-    ode.paths.METADATA_PATH = (tmp_path / ".metadata")
-    return tmp_path
 
 
 class TestPaths:
